+++ server/lib/class/questionnaires.js
@@ -53,7 +53,7 @@ function Questionnaires( ) {
 	 */
 	this.getQuestionnaires = function(pg, offset) {
 		logger.trace("getQuestionnaires");
-		var cursor = physioDOM.db.collection("questionnaires").find({}, {name: 1, text: 1});
+		var cursor = physioDOM.db.collection("questionnaires").find({}, {name: 1, ref: 1, label: 1});
 		cursor = cursor.sort( {name: 1} );
 		return dbPromise.getList(cursor, pg, offset);
 	};
@@ -64,7 +64,7 @@ function Questionnaires( ) {
 	 * @param qName
 	 * @returns {promise}
 	 */
-	this.getQuestionnaire = function(qName) {
+	this.getQuestionnaireByName = function(qName) {
 		var that = this;
 		return new promise( function(resolve, reject) {
 			logger.trace("getQuestionnaire", qName);
@@ -85,6 +85,19 @@ function Questionnaires( ) {
 				}
 			});
 		});
+	};
+
+	/**
+	 * get a questionnaire by its id
+	 *
+	 * @param qName
+	 * @returns {promise}
+	 */
+	this.getQuestionnaireByID = function(entryID) {
+		logger.trace("getBeneficiaryByID", entryID);
+		var questionnaireByID = new ObjectID(entryID);
+		var questionnaire = new Questionnaire();
+		return questionnaire.getById(questionnaireByID);
 	};
 }